Move ServiceInstances trait
Move the ServiceInstances trait to the service::network module.  This
will decouple the trait from the handler module.
