+++ libsplinter/src/service/network/mod.rs
@@ -23,6 +23,7 @@ use std::collections::{BTreeMap, HashMap};
 use std::sync::mpsc::{self, Receiver, Sender};
 use std::thread;
 
+use crate::circuit::service::ServiceId;
 use crate::network::connection_manager::{ConnectionManagerNotification, Connector};
 
 use self::error::ServiceConnectionAgentError;
@@ -83,6 +84,36 @@ pub enum ServiceConnectionNotification {
         service_id: String,
         endpoint: String,
     },
+}
+
+/// A mapping of service instances and the component responsible for it.  This can be used to add
+/// or remove service connection information.
+pub trait ServiceInstances {
+    /// Add a service instance.
+    ///
+    /// This method should create an association of the service with the given component id.
+    ///
+    /// # Errors
+    ///
+    /// Returns a `ServiceAddInstanceError` if the service cannot be added.
+    fn add_service_instance(
+        &self,
+        service_id: ServiceId,
+        component_id: String,
+    ) -> Result<(), ServiceAddInstanceError>;
+
+    /// Remove a service instance.
+    ///
+    /// This method should remove the association of the service with the given component id.
+    ///
+    /// # Errors
+    ///
+    /// Returns a `ServiceRemoveInstanceError` if the service cannot be removed.
+    fn remove_service_instance(
+        &self,
+        service_id: ServiceId,
+        component_id: String,
+    ) -> Result<(), ServiceRemoveInstanceError>;
 }
 
 /// Constructs new ServiceConnectionManager structs.
